# local-shared-data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/webhook/xhs-result": {
      "post": {
        "tags": [
          "webhook"
        ],
        "summary": "Receive Xhs Webhook",
        "description": "\u63a5\u6536\u5c0f\u7ea2\u4e66\u722c\u866b\u7ed3\u679c\u7684webhook\u7aef\u70b9",
        "operationId": "receive_xhs_webhook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebhookRequest": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/XhsSearchResult"
              },
              {
                "$ref": "#/components/schemas/XhsNoteData"
              },
              {
                "$ref": "#/components/schemas/XhsUserResult"
              },
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "status": {
            "$ref": "#/components/schemas/WebhookStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "elapsed_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Time"
          },
          "progress": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message",
          "timestamp",
          "run_id"
        ],
        "title": "WebhookRequest",
        "description": "\u5b8c\u6574\u7684Webhook\u8bf7\u6c42\u6a21\u5f0f"
      },
      "WebhookResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "default": "received"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "\u6570\u636e\u5df2\u63a5\u6536"
          },
          "processed_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed Count"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "type": "object",
        "title": "WebhookResponse",
        "description": "Webhook\u54cd\u5e94"
      },
      "WebhookStatus": {
        "type": "string",
        "enum": [
          "started",
          "progress",
          "success",
          "error",
          "completed",
          "failed"
        ],
        "title": "WebhookStatus",
        "description": "Webhook\u72b6\u6001\u679a\u4e3e"
      },
      "XhsAuthorInfo": {
        "properties": {
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          }
        },
        "type": "object",
        "title": "XhsAuthorInfo",
        "description": "\u5c0f\u7ea2\u4e66\u4f5c\u8005\u4fe1\u606f"
      },
      "XhsInteractInfo": {
        "properties": {
          "liked_count": {
            "type": "integer",
            "title": "Liked Count",
            "default": 0
          },
          "collected_count": {
            "type": "integer",
            "title": "Collected Count",
            "default": 0
          },
          "comment_count": {
            "type": "integer",
            "title": "Comment Count",
            "default": 0
          },
          "share_count": {
            "type": "integer",
            "title": "Share Count",
            "default": 0
          }
        },
        "type": "object",
        "title": "XhsInteractInfo",
        "description": "\u5c0f\u7ea2\u4e66\u4e92\u52a8\u4fe1\u606f"
      },
      "XhsNoteData": {
        "properties": {
          "note_id": {
            "type": "string",
            "title": "Note Id"
          },
          "note_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Url"
          },
          "note_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Type"
          },
          "author": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/XhsAuthorInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "desc": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desc"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "upload_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upload Time"
          },
          "ip_location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Location"
          },
          "interact_info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/XhsInteractInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "video_cover": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Cover"
          },
          "video_addr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Addr"
          },
          "image_list": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image List"
          },
          "xsec_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Xsec Token"
          }
        },
        "type": "object",
        "required": [
          "note_id"
        ],
        "title": "XhsNoteData",
        "description": "\u5c0f\u7ea2\u4e66\u7b14\u8bb0\u6570\u636e"
      },
...
      "XhsSearchResult": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "total_found": {
            "type": "integer",
            "title": "Total Found"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/XhsNoteData"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "query",
          "total_found",
          "notes"
        ],
        "title": "XhsSearchResult",
        "description": "\u641c\u7d22\u7ed3\u679c\u6570\u636e"
      },
      "XhsUserResult": {
        "properties": {
          "user_url": {
            "type": "string",
            "title": "User Url"
          },
          "notes_count": {
            "type": "integer",
            "title": "Notes Count"
          },
          "notes": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "user_url",
          "notes_count",
          "notes"
        ],
        "title": "XhsUserResult",
        "description": "\u7528\u6237\u7ed3\u679c\u6570\u636e"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  }
}